requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web App dependencies
flask>=2.3.0
groq>=0.4.0
google-generativeai>=0.3.0
mysql-connector-python>=8.0.33

